refactor(profile): centralize Profile interface to shared types file
- Move Profile interface to lib/types/profile.ts
- Update all profile components to import from centralized location
- Eliminate interface duplication across multiple files
- Improve maintainability and type consistency
