+++ components/profile/sections/about-section.tsx
@@ -1,7 +1,7 @@
 import { FiMapPin, FiBriefcase, FiGlobe, FiCalendar } from "react-icons/fi";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { ProfileTag } from "../profile-tag";
-import type { Profile } from "../profile-content";
+import type { Profile } from "@/lib/types/profile";
 
 interface AboutSectionProps {
   profile: Profile;
